moved config to single domain common for weather and sensor platforms. Code simplified
